requirements.txt
flask
flask_cors
flask_restful
flask_sqlalchemy
Flask-Security-Too
email_validator
bcrypt
python-dotenv
uagents
langchain
langchain_openai
flask_jwt_extended
flask_bcrypt
tzlocal
twilio
pandas